Add isEmpty method to createSquare
Added the isEmpty method to createSquare that will be used to check
if a square is empty before executing a move.

It's important to note that Gameboard currently lacks validation for
"nonsense" moves, such as marking a square with non-integer or out-of-
bounds indices. As the user interface dictates the details of move
input, the responsibility for validation of such moves is delegated to
the user interface layer.