Add GameController module for managing game flow
Added the GameController module that controls the flow of the game, e.g.
playing a turn, switching the current player, checking for a win or draw
etc.

Note that the module currently contains logic for printing the player's
turn, move, and more, for testing purposes. This will shortly be moved
to the user interface layer.