Modify getBoard method in Gameboard and refactor
- Modified the getBoard method in Gameboard to return a copy of the
  board with the values of the squares instead of the square objects.
  This encapsulates the game board by preventing external code from
  accessing the board directly
- Refactored the code in accordance with these changes

NOTE: Returning a copy of the board with values instead of objects means
the isEmpty method can't be used to check if a square is empty. The
alternative to this would be to return a deep copy of the game board,
but that would hurt the performance as a side effect.